+++ tests/test_common_metadata.py
@@ -176,6 +176,8 @@ class CommonMetadataTest(unittest.TestCase):
         x.common_metadata.description = example_description
         self.assertEqual(x.common_metadata.description, example_description)
         self.assertEqual(x.properties["description"], example_description)
+        with self.assertRaises(ValueError):
+            x.common_metadata.description = ""
 
         # License
         license = "PDDL-1.0"
@@ -538,3 +540,47 @@ class AssetCommonMetadataTest(unittest.TestCase):
         self.assertEqual(
             analytic.to_dict()["updated"], utils.datetime_to_str(set_value)
         )
+
+    def test_keywords(self) -> None:
+        item = self.item.clone()
+        cm = item.common_metadata
+        analytic = item.assets["analytic"]
+        analytic_cm = CommonMetadata(analytic)
+        thumbnail = item.assets["thumbnail"]
+        thumbnail_cm = CommonMetadata(thumbnail)
+
+        item_value = cm.keywords
+        a2_known_value = ["keyword_a"]
+
+        # Get
+        self.assertNotEqual(thumbnail_cm.keywords, item_value)
+        self.assertEqual(thumbnail_cm.keywords, a2_known_value)
+
+        # Set
+        set_value = ["keyword_b"]
+        analytic_cm.keywords = set_value
+
+        self.assertEqual(analytic_cm.keywords, set_value)
+        self.assertEqual(analytic.to_dict()["keywords"], set_value)
+
+    def test_roles(self) -> None:
+        item = self.item.clone()
+        cm = item.common_metadata
+        analytic = item.assets["analytic"]
+        analytic_cm = CommonMetadata(analytic)
+        thumbnail = item.assets["thumbnail"]
+        thumbnail_cm = CommonMetadata(thumbnail)
+
+        item_value = cm.roles
+        a2_known_value = ["a_role"]
+
+        # Get
+        self.assertNotEqual(thumbnail_cm.roles, item_value)
+        self.assertEqual(thumbnail_cm.roles, a2_known_value)
+
+        # Set
+        set_value = ["another_role"]
+        analytic_cm.roles = set_value
+
+        self.assertEqual(analytic_cm.roles, set_value)
+        self.assertEqual(analytic.to_dict()["roles"], set_value)